deployment/requirements.txt
# Cleansheet Deployment Tool Dependencies
# Python 3.10+

# CLI framework
click>=8.1.0

# Beautiful terminal output
rich>=13.0.0

# Configuration file handling
pyyaml>=6.0

# Testing framework
pytest>=7.0.0

# Additional utilities
python-dateutil>=2.8.0
